Move subset of source-map-consumer.js to basic-source-map-consumer.js
